[#125] разбивка проектов на архивные и активные

+++ app/decorators/membership_decorator.rb
@@ -5,7 +5,7 @@ class MembershipDecorator < Draper::Decorator
   delegate :name, to: :user
 
   def remove_link
-    if h.current_user.can_delete?(source)
+    if source.project.active and h.current_user.can_delete?(source)
       h.link_to h.project_membership_path(source.project, source), :data => { method: :delete, confirm: I18n.t('memberships.delete.confirm') } do
         h.ficon 'cancel-1', color: :gray, size: 18
       end
